Revert "Extract hub and simplify status update logic (#49)"
This reverts commit f838e398b057b1c2790987d11c9d3e8fc8de1384.
